refactor(mem_cache): split tree_node_lru into components, rename LRUSlot->Slot
Move the Slot trait + shared structs (EvictRequest/EvictResult/
ComponentPoolState/LRUData) to components/mod.rs, per-component Slot impls
to full/swa/mamba.rs, and eviction helpers to components/utils.rs; delete
tree_node_lru.rs. Drop the LRU prefix from the slot types (they own values
+ lock_ref + eviction, not just LRU); keep LRUData (the intrusive list node).
